requirements.txt
numpy
pandas
matplotlib
plotly-express
pandas_datareader
yfinance
transformers 
torch 
streamlit